+++ src/engine/z_buffer.c
@@ -13,7 +13,7 @@ void	fill_z_buff(float *z_buffer, size_t length)
 	}
 }
 
-float	*get_z_buffer_value(float *z_buffer, size_t x, size_t y, size_t width)
+float	*get_z_buffer_value(float *z_buffer, int x, int y, int width)
 {
 	return (z_buffer + y * width + x);
 }
